// src/app/interfaces/IDobleAsesoria.ts
export interface IDobleAsesoria {
    persistenceId: number;
    persistenceId_string: string;
    persistenceVersion: number;
    persistenceVersion_string: string;
    identificacion: string;
    nombres?: any;
    apellidos?: any;
    aliasAsesor?: any;
    fecha?: any;
    tipoGestion?: any;
    fechaNacimiento?: any;
    regimen?: any;
    tiempoAfiliacion?: any;
    sexo?: any;
    respuestaCliente?: any;
    resultadoViabilidad?: any;
    resultadoDobleAsesoria?: any;
    alternativaSeleccionada?: any;
    solicitudInmediata?: any;
    solicitudRegistrada?: any;
    email?: any;
    direccion?: any;
    telefono?: any;
    fechaOportunidad?: any;
    salario?: any;
    fechaVisita?: any;
}